Changed and broke a lot

+++ PhoneShop.Business/Extensions/PhoneExtensions.cs
@@ -4,7 +4,7 @@ namespace Phoneshop.Business.Extensions
 {
     public static class PhoneExtensions
     {
-        public static double PriceWithoutVat(this Phone value)
+        public static decimal PriceWithoutVat(this Phone value)
         {
             return value.Price - value.Price / 100 * 21;
         }
